--- vips/vip-382/bscmainnet.ts
@@ -1,29 +1,0 @@
-import { ProposalType } from "src/types";
-import { makeProposal } from "src/utils";
-import { FAST_TRACK_TIMELOCK } from "src/vip-framework";
-
-const vip382 = () => {
-  const meta = {
-    version: "v2",
-    title: "VIP-382 [zkSync] Reduce the distribution speeds of the markets on zkSync",
-    description: ``,
-    forDescription: "I agree that Venus Protocol should proceed with this proposal",
-    againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
-    abstainDescription: "I am indifferent to whether Venus Protocol proceeds or not",
-  };
-
-  return makeProposal(
-    [
-      {
-        target: FAST_TRACK_TIMELOCK,
-        signature: "",
-        params: [],
-        value: "1",
-      },
-    ],
-    meta,
-    ProposalType.FAST_TRACK,
-  );
-};
-
-export default vip382;
